# Conflicts:
#	dpp/src/data_contract/extra/document_field.rs
#	dpp/src/data_contract/extra/index.rs
#	dpp/src/document/state_transition/documents_batch_transition/validation/basic/find_duplicates_by_indices.rs
#	dpp/src/errors/codes.rs
#	dpp/src/errors/consensus/abstract_consensus_error.rs
#	dpp/src/errors/consensus/basic/abstract_basic_error.rs
#	dpp/src/errors/consensus/basic/identity/invalid_instant_asset_lock_proof_signature_error.rs
#	dpp/src/errors/consensus/basic/identity/missing_master_public_key_error.rs
#	dpp/src/errors/consensus/basic/identity/mod.rs
#	dpp/src/errors/non_consensus_error.rs
#	dpp/src/errors/serde_parsing_error.rs
#	dpp/src/identity/errors/asset_lock_output_not_found_error.rs
#	dpp/src/identity/identity.rs
#	dpp/src/identity/state_transition/asset_lock_proof/asset_lock_proof_validator.rs
#	dpp/src/identity/state_transition/asset_lock_proof/asset_lock_public_key_hash_fetcher.rs
#	dpp/src/identity/state_transition/asset_lock_proof/asset_lock_transaction_output_fetcher.rs
#	dpp/src/identity/state_transition/asset_lock_proof/asset_lock_transaction_validator.rs
#	dpp/src/identity/state_transition/asset_lock_proof/chain/chain_asset_lock_proof.rs
#	dpp/src/identity/state_transition/asset_lock_proof/chain/mod.rs
#	dpp/src/identity/state_transition/asset_lock_proof/instant/instant_asset_lock_proof.rs
#	dpp/src/identity/state_transition/asset_lock_proof/instant/instant_asset_lock_proof_structure_validator.rs
#	dpp/src/identity/state_transition/asset_lock_proof/mod.rs
#	dpp/src/identity/state_transition/identity_create_transition/apply_identity_create_transition.rs
#	dpp/src/identity/state_transition/identity_create_transition/identity_create_transition.rs
#	dpp/src/identity/state_transition/identity_create_transition/mod.rs
#	dpp/src/identity/state_transition/identity_create_transition/validation/basic/identity_create_transition_basic_validator.rs
#	dpp/src/identity/state_transition/identity_create_transition/validation/basic/mod.rs
#	dpp/src/identity/state_transition/identity_create_transition/validation/mod.rs
#	dpp/src/identity/state_transition/identity_topup_transition/apply_identity_topup_transition.rs
#	dpp/src/identity/state_transition/identity_topup_transition/identity_topup_transition.rs
#	dpp/src/identity/state_transition/identity_topup_transition/mod.rs
#	dpp/src/identity/state_transition/identity_topup_transition/validation/basic/identity_topup_transition_basic_validator.rs
#	dpp/src/identity/state_transition/identity_topup_transition/validation/mod.rs
#	dpp/src/identity/state_transition/mod.rs
#	dpp/src/identity/validation/public_keys_in_identity_transition_validator.rs
#	dpp/src/mocks.rs
#	dpp/src/state_repository.rs
#	dpp/src/state_transition/mod.rs
#	dpp/src/tests/fixtures/get_documents_fixture.rs
#	dpp/src/tests/fixtures/instant_asset_lock_proof_fixture.rs
#	dpp/src/tests/fixtures/mod.rs
#	dpp/src/tests/identity/state_transition/asset_lock/instant/mod.rs
#	dpp/src/tests/identity/state_transition/identity_create_transition/validation/basic/identity_create_transition_basic_validator_spec.rs
#	dpp/src/tests/identity/state_transition/mod.rs
#	dpp/src/tests/utils/utils.rs
#	js-binding/src/identity.rs

+++ dpp/src/identity/state_transition/identity_create_transition/apply_identity_create_transition.rs
@@ -1,12 +1,13 @@
+use std::sync::Arc;
+
+use anyhow::{anyhow, Result};
+
 use crate::identity::state_transition::asset_lock_proof::AssetLockTransactionOutputFetcher;
 use crate::identity::state_transition::identity_create_transition::IdentityCreateTransition;
 use crate::identity::{convert_satoshi_to_credits, Identity};
-use crate::ProtocolError;
-use anyhow::{anyhow, Result};
-use std::sync::Arc;
-
 use crate::state_repository::StateRepositoryLike;
 use crate::state_transition::StateTransitionLike;
+use crate::ProtocolError;
 
 pub struct ApplyIdentityCreateTransition<SR>
 where
@@ -29,7 +30,7 @@ where
             .fetch(state_transition.get_asset_lock_proof())
             .await?;
 
-        let credits_amount = convert_satoshi_to_credits(output.value as i64);
+        let credits_amount = convert_satoshi_to_credits(output.value);
 
         let identity = Identity {
             protocol_version: state_transition.get_protocol_version(),
